requirements.txt
pyfiglet==0.8.post1
python-nmap==0.7.1

